ble_gateway/requirements.txt
bleak
PyYAML